+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [1.0.0-beta.1] - 2020-01-19
+
 ### Added
 - Added ReactDOM-like prop validation in development ([#47])
 
@@ -350,7 +353,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `<TilingSprite />` component
 
 
-[Unreleased]: https://github.com/michalochman/react-pixi-fiber/compare/v0.13.0...HEAD
+[Unreleased]: https://github.com/michalochman/react-pixi-fiber/compare/v1.0.0-beta.1...HEAD
+[1.0.0-beta.1]: https://github.com/michalochman/react-pixi-fiber/compare/v0.13.0...v1.0.0-beta.1
 [0.13.0]: https://github.com/michalochman/react-pixi-fiber/compare/v0.12.3...v0.13.0
 [0.12.3]: https://github.com/michalochman/react-pixi-fiber/compare/v0.12.2...v0.12.3
 [0.12.2]: https://github.com/michalochman/react-pixi-fiber/compare/v0.12.1...v0.12.2